fix: export setLayerOpacity method from storymap plugin
