+++ .changeset/green-horses-fail.md
@@ -1,0 +1,5 @@
+---
+"@undp-data/svelte-maplibre-storymap": patch
+---
+
+fix: export setLayerOpacity method from storymap plugin
