+++ packages/svelte-maplibre-storymap/src/lib/index.ts
@@ -4,3 +4,4 @@ import StoryMapChapter from './StoryMapChapter.svelte';
 export { StoryMap, StoryMapChapter };
 export * from './interfaces/index.js';
 export * from './stores/index.js';
+export * from './helpers.js';
